requirements.txt
cvxpy==1.8.1
matplotlib==3.10.7
numpy==2.4.2
pandas==3.0.1
python-dotenv==1.2.2
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.1
typer==0.24.1
pytest==8.4.2
